+++ actix-utils/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Minimum supported Rust version (MSRV) is now 1.65.
+- Minimum supported Rust version (MSRV) is now 1.70.
 
 ## 3.0.1