+++ actix-utils/CHANGES.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
+## 3.0.2
+
+- Deprecate `ready` in favor of `core::future::{ready, Ready}`.
 - Minimum supported Rust version (MSRV) is now 1.88.
-- Deprecate `crate::ready::*` in favor of `core::future::{ready, Ready}`
 
 ## 3.0.1